Added JWT claims

+++ resources/item.py
@@ -7,7 +7,7 @@ Created on Sun Feb  2 23:46:24 2020
 
 from flask import request
 from flask_restful import Resource
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 from models.item import ItemModel
 
 class Item(Resource):
@@ -21,14 +21,17 @@ class Item(Resource):
   
     def post(self, name):
         if ItemModel.find_by_name(name):
-            return {"message":"An item with the same name already exists"}, 400
+            return {"message":"An item with the same name already exists"}, 401
         data = request.get_json()
         item=ItemModel(name,data['price'],data['store_id'])
         item.save_to_db()
         return item.json()
 
-    
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims["is admin"]:
+            return {"message":"Admin priviledge required"}, 401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
